+++ client/src/AuthContext.jsx
@@ -100,7 +100,7 @@ export const AuthProvider = ({ children }) => {
     setUser(null);
     delete axios.defaults.headers.common["Authorization"];
     localStorage.removeItem("access_token");
-    router.push("/");
+    router("/");
   };
 
   return (
