--- client/src/components/AuthForm/AuthForm.css
@@ -1,3 +1,0 @@
-.button_container {
-  gap: 10px;
-}
