+++ client/src/components/AuthForms/AuthForms.css
@@ -1,0 +1,98 @@
+.auth-form-page {
+  max-width: 460px;
+  margin: 28px auto;
+  padding: 0 12px;
+}
+
+.auth-form-card {
+  display: flex;
+  flex-direction: column;
+  gap: 14px;
+  background: #fff;
+  border-radius: 10px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  padding: 24px;
+}
+
+.auth-form-card legend {
+  font-size: 24px;
+  font-weight: 700;
+  color: #333;
+}
+
+.auth-form-subtitle {
+  margin: 0;
+  color: #555;
+}
+
+.auth-form-card fieldset {
+  border: none;
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.auth-form-card label {
+  font-weight: 600;
+  color: #333;
+}
+
+.auth-form-card input {
+  width: 100%;
+  box-sizing: border-box;
+  padding: 10px 12px;
+  border: 1px solid #ddd;
+  border-radius: 6px;
+  font-size: 14px;
+  font-family: inherit;
+}
+
+.auth-form-card input:focus {
+  outline: none;
+  border-color: #007bff;
+}
+
+.button_container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+  width: 100%;
+}
+
+.button_container button {
+  border: none;
+  border-radius: 6px;
+  padding: 10px 14px;
+  font-size: 14px;
+  font-weight: 600;
+  cursor: pointer;
+  background: #007bff;
+  color: #fff;
+  transition: background-color 0.2s ease;
+}
+
+.button_container .redirect_button {
+  background: #f5f7fb;
+  color: #1f4fa3;
+  border: 1px solid #b8c9ea;
+}
+
+.button_container button:hover {
+  background: #0056b3;
+}
+
+.button_container .redirect_button:hover {
+  background: #e9f0ff;
+}
+
+.button_container a {
+  text-decoration: none;
+  order: 1;
+}
+
+.button_container button[type="submit"] {
+  order: 2;
+}
