--- client/src/pages/AuthPage.jsx
@@ -1,11 +1,0 @@
-import AuthForm from "../components/AuthForm/AuthForm.jsx";
-
-const AuthPage = () => {
-  return (
-    <div>
-      <AuthForm />
-    </div>
-  );
-};
-
-export default AuthPage;
